audio: base_fw: move L1_EXIT control to platform specific code
Move the Intel specific L1_EXIT control handling from base_fw.c
to Intel specific platform code.
